Reduce UI delay from backend trips
Now we just modify the local state and hope that 1) we're paying
attention and doing the state transitions right and 2) the backend never
fails.

It would be nice to add a state rollback feature in the case that the
backend errors out on a request, so the UI can update immediately but
roll back if an error occurs.